refactor(renderer): split renderer-app.tsx by responsibility (Phase 6)
Extract three focused modules from renderer-app.tsx to bring it under
600 LOC and separate concerns cleanly:

- src/renderer/ipc-listeners.ts (46 LOC): wireIpcListeners /
  unwireIpcListeners — isolates all IPC subscription management.
- src/renderer/settings-mutations.ts (452 LOC): createSettingsMutations
  factory — all settings, preset, and API-key mutation helpers bound to
  injected deps.
- src/renderer/native-recording.ts (393 LOC): recording lifecycle
  (start/stop/cancel/toggle), audio-source discovery, and
  handleRecordingCommandDispatch.

renderer-app.tsx shrinks from 1275 → 581 LOC; it is now the thin
orchestration layer covering boot, state, autosave, and render wiring.

Public exports (startRendererApp, stopRendererAppForTests) unchanged.
All 45 renderer tests pass. Codex review found no behavioral regressions.
